+++ jmri/jmrix/easydcc/EasyDccOpsModeProgrammer.java
@@ -12,7 +12,7 @@ import jmri.*;
  *
  * @see             jmri.Programmer
  * @author			Bob Jacobsen Copyright (C) 2002
- * @version			$Revision: 1.5 $
+ * @version			$Revision: 1.6 $
  */
 public class EasyDccOpsModeProgrammer extends EasyDccProgrammer  {
 
@@ -39,7 +39,7 @@ public class EasyDccOpsModeProgrammer extends EasyDccProgrammer  {
         int j = 4;
         for (int i=0; i<contents.length; i++) {
             msg.setElement(j++, ' ');
-            msg.addIntAsTwoHex(((int)contents[i])&0xFF, j);
+            msg.addIntAsTwoHex(contents[i]&0xFF, j);
             j = j+2;
         }
